chore: restructure tests [wip]
